refactor(desktop): organize V2 project settings into titled sections
Groups the flat settings list into General, Branches & naming,
Location & checkout, Scripts, and Danger zone. Promotes SettingsSection
from ModelsSettings to the shared settings components (used by two
pages now) and bumps its title to text-base font-semibold so section
headers read above row labels.
