+++ apps/desktop/src/renderer/routes/_authenticated/settings/models/components/ModelsSettings/ModelsSettings.tsx
@@ -17,13 +17,13 @@ import { AnthropicOAuthDialog } from "renderer/components/Chat/ChatInterface/com
 import { OpenAIOAuthDialog } from "renderer/components/Chat/ChatInterface/components/ModelPicker/components/OpenAIOAuthDialog";
 import { useAnthropicOAuth } from "renderer/components/Chat/ChatInterface/components/ModelPicker/hooks/useAnthropicOAuth";
 import { useOpenAIOAuth } from "renderer/components/Chat/ChatInterface/components/ModelPicker/hooks/useOpenAIOAuth";
+import { SettingsSection } from "renderer/routes/_authenticated/settings/components/SettingsSection";
 import {
 	isItemVisible,
 	SETTING_ITEM_ID,
 	type SettingItemId,
 } from "../../../utils/settings-search";
 import { ConfigRow } from "./components/ConfigRow";
-import { SettingsSection } from "./components/SettingsSection";
 import {
 	buildAnthropicEnvText,
 	EMPTY_ANTHROPIC_FORM,
